refactor: decompose conventions.rs god file into focused modules
Extract 5 modules from conventions.rs (1,966 → ~1,100 lines):
- discovery.rs: auto_discover_groups, discover_cross_directory + 8 tests
- fingerprint.rs: FileFingerprint struct + fingerprint_file
- import_matching.rs: has_import, grouped_import_contains + 9 tests
- signatures.rs: normalize/tokenize/compute_signature_skeleton
- walker.rs: walk_source_files, count_unclaimed, extension helpers

Add test_helpers.rs with shared make_convention() factory for test
modules that construct ConventionReport instances.

Closes #328
